Move relevant figure scripts to subdirectory
